Move and rename "skstore*" TS sources to "skruntime-ts"
