+++ skruntime-ts/ts/src/internals/skruntime_impl.ts
@@ -1,6 +1,6 @@
 // prettier-ignore
 import { type ptr, type Opt } from "#std/sk_types.js";
-import type { Context } from "./skstore_types.js";
+import type { Context } from "./skruntime_types.js";
 import type * as Internal from "./skstore_internal_types.js";
 import type {
   Accumulator,
@@ -24,7 +24,7 @@ import type {
   LazyCompute,
   AsyncLazyCompute,
   NonEmptyIterator,
-} from "../skstore_api.js";
+} from "../skip_runtime_api.js";
 
 // prettier-ignore
 import type { MirrorDefn, Params, SKDBSync } from "#skdb/skdb_types.js";
